requirements.txt
absl-py==1.1.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.2.0
alembic==1.13.1
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
autopage==0.5.2
Bottleneck==1.3.4
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.0
cfgv==3.4.0
charset-normalizer==2.0.4
chumpy==0.70
click==8.1.7
cliff==4.7.0
cloudpickle==1.6.0
cmaes==0.10.0
cmd2==2.4.3
colorama==0.4.4
coloredlogs==15.0.1
colorlog==6.8.2
conda-package-handling==2.2.0
conda_package_streaming==0.9.0
contourpy==1.0.7
cryptography==36.0.0
cycler==0.10.0
Cython==0.29.28
cytoolz==0.11.0
dask==2.30.0
decorator==4.4.2
distlib==0.3.8
docker-pycreds==0.4.0
easydict==1.9
einops==0.8.0
exceptiongroup==1.2.1
faiss==1.7.3
filelock==3.14.0
flatbuffers==2.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2024.5.0
ftfy==6.2.0
fvcore==0.1.5.post20210915
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.9.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.8.4
greenlet==3.0.3
grpcio==1.47.0
h5py==3.8.0
huggingface-hub==0.22.2
humanfriendly==10.0
hydra-colorlog==1.2.0
hydra-core==1.2.0
hydra-optuna-sweeper==1.2.0
identify==2.5.36
idna==3.3
imageio==2.9.0
importlib-metadata==4.12.0
importlib-resources==5.11.0
iniconfig==2.0.0
insightface==0.6.2
iopath==0.1.9
joblib==1.2.0
keras==2.11.0
kiwisolver==1.2.0
libclang==15.0.6.1
lightning-utilities==0.8.0
lmdb==1.2.1
loguru==0.6.0
lpips==0.1.4
Mako==1.3.5
Markdown==3.3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.0
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.2.1
multidict==6.0.4
mxnet==1.9.1
networkx==2.5
nodeenv==1.8.0
numexpr==2.8.1
numpy==1.24.2
oauthlib==3.2.0
omegaconf==2.3.0
onnx==1.12.0
onnxruntime==1.11.0
onnxruntime-gpu==1.13.1
opencv-python==4.7.0.68
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
optuna==2.10.1
packaging==21.3
pandas==2.0.0
pbr==6.0.0
Pillow==9.4.0
pip==22.3.1
platformdirs==4.2.2
plot-model==0.20
pluggy==1.5.0
portalocker==2.4.0
pre-commit==3.5.0
prettytable==3.3.0
protobuf==3.19.4
provider==0.1.0
psbody-mesh==0.4
psutil==5.9.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.2
Pygments==2.18.0
PyOpenGL==3.1.6
pyOpenSSL==22.0.0
pyparsing==2.4.7
pyperclip==1.8.2
pyrootutils==1.0.4
PySocks==1.7.1
pytest==8.2.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytorch-fid==0.2.0
pytorch-ignite==0.4.11
pytorch-lightning==1.7.1
pytorch3d==0.7.1
pytz==2021.3
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==25.0.0
qudida==0.0.4
regex==2024.5.15
requests==2.27.1
requests-oauthlib==1.3.1
rich==13.7.1
rsa==4.8
scikit-image==0.17.2
scikit-learn==1.2.2
scipy==1.10.0
seaborn==0.12.2
sentry-sdk==2.3.1
setproctitle==1.3.3
setuptools==67.3.2
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.30
stevedore==5.2.0
sympy==1.11.1
tabulate==0.8.9
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2020.10.1
tomli==2.0.1
toolz==0.11.1
torch==1.11.0
torchaudio==0.11.0
torchmetrics==0.9.3
torchvision==0.12.0
tornado==6.1
tqdm==4.63.1
trimesh==3.12.7
tsnecuda==3.0.1+cu112
typing_extensions==4.12.0
tzdata==2024.1
UNKNOWN==0.0.0
urllib3==1.26.18
virtualenv==20.26.2
wandb==0.17.0
wcwidth==0.2.13
Werkzeug==1.0.1
wheel==0.37.1
wrapt==1.15.0
yacs==0.1.8
yarl==1.8.2
zipp==3.8.0
zstandard==0.21.0
